refactor: rename everything to ts, fix imports and exports
