+++ src/app/modules/verification/verification.factory.ts
@@ -1,6 +1,8 @@
-const featureManager = require('../../../config/feature-manager').default
-const verification = require('./verification.server.controller')
-const HttpStatus = require('http-status-codes')
+import HttpStatus from 'http-status-codes'
+
+import featureManager from '../../../config/feature-manager'
+
+import * as verification from './verification.controller'
 
 const verifiedFieldsFactory = ({ isEnabled }) => {
   if (isEnabled) {
@@ -28,4 +30,4 @@ const verifiedFieldsFactory = ({ isEnabled }) => {
   }
 }
 
-module.exports = verifiedFieldsFactory(featureManager.get('verified-fields'))
+export default verifiedFieldsFactory(featureManager.get('verified-fields'))
