+++ src/loaders/express/index.ts
@@ -8,7 +8,7 @@ import path from 'path'
 import url from 'url'
 
 import mountSnsRoutes from '../../app/modules/sns/sns.routes'
-import mountVfnRoutes from '../../app/modules/verification/verification.server.routes'
+import mountVfnRoutes from '../../app/modules/verification/verification.routes'
 import apiRoutes from '../../app/routes'
 import config from '../../config/config'
